Example RNA-Seq data from paired observations

Sample RescueSimParams object with gene expression data for recruited airspace macrophage cells from bronchoalveolar lavage samples samples collected from healthy adults. The dataset from which parameters were estimated had 19,410 genes and 9,041 cells. Data was collected from five subjects at two timepoints per subject.

Usage

data(RecAM_params)

Format

RescueSimParams object with parameters estimated from empirical dataset