Simulate Repeated Measures scRNA-seq Data

Simulate Repeated Measures scRNA-seq Data using the RESCUE method

simRescueData(paramObj)

Arguments

paramObj

RescueSimParams-class object with no empty slots.

Value

SingleCellExperiment object with the data in the following slots

counts

Matrix of raw counts with genes represented by rows and cells represented by columns.

colData

sampleID

Sample identifier

subjectID

Subject identifier

time

Timepoint identifier

group

Group identifier

rowData

deLog2FC

A DataFrame containing log2 fold change information for each gene. Each column corresponds to a non-reference experimental condition (e.g., "time1", "group1", "time1_group1", etc.), and values represent gene-level log2 fold changes relative to the baseline condition ("time0", "group0", or "time0_group0" depending on the design). The reference condition itself is not included as a column.

Examples

# Read in data
 data("RecAM_sce")

 # Calculate sim parameters for first 50 genes
RecAM_sce <- RecAM_sce[1:50,]
RecAM_params<-estRescueSimParams(RecAM_sce, sampleVariable = "sampleID",
subjectVariable = "subjectID", timepointVariable = "time")

# Simulate data
simDat=simRescueData(RecAM_params)

# Examine
library(SingleCellExperiment)
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> Loading required package: generics
#> 
#> Attaching package: ‘generics’
#> The following objects are masked from ‘package:base’:
#> 
#>     as.difftime, as.factor, as.ordered, intersect, is.element, setdiff,
#>     setequal, union
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, is.unsorted, lapply,
#>     mapply, match, mget, order, paste, pmax, pmax.int, pmin, pmin.int,
#>     rank, rbind, rownames, sapply, saveRDS, table, tapply, unique,
#>     unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
counts(simDat)[1:5, 1:5]
#>            cell_1 cell_2 cell_3 cell_4 cell_5
#> NOC2L           0      0      1      0      1
#> SCNN1D          0      0      0      0      0
#> AL391244.3      0      0      0      0      0
#> AL645728.1      0      0      0      0      0
#> AL691432.2      0      0      0      0      0